Q q Q
q

+++ src/Main.java
@@ -1,7 +1,7 @@
 public class Main {
     public static void main(String[] args) {
-       SalesManager salesManager = new SalesManager(new int[] {15, 100, 500, 150});
-       int maxSale = salesManager.max();
+       SalesManager salesManager = new SalesManager(new long[] {15, 100, 500, 150});
+        long maxSale = salesManager.max();
         System.out.println(maxSale);
